@@ -1,32 +1,55 @@
 <?php
 
+use Illuminate\Support\Facades\Route;
+
+use App\Http\Controllers\CardController;
+use App\Http\Controllers\ItemController;
+
+use App\Http\Controllers\Auth\LoginController;
+use App\Http\Controllers\Auth\RegisterController;
+
 /*
 |--------------------------------------------------------------------------
 | Web Routes
 |--------------------------------------------------------------------------
 |
 | Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
+| routes are loaded by the RouteServiceProvider and all of them will
+| be assigned to the "web" middleware group. Make something great!
 |
 */
+
 // Home
-Route::get('/', 'Auth\LoginController@home');
+Route::redirect('/', '/login');
 
 // Cards
-Route::get('cards', 'CardController@list');
-Route::get('cards/{id}', 'CardController@show');
+Route::controller(CardController::class)->group(function () {
+    Route::get('/cards', 'list')->name('cards');
+    Route::get('/cards/{id}', 'show');
+});
+
 
 // API
-Route::put('api/cards', 'CardController@create');
-Route::delete('api/cards/{card_id}', 'CardController@delete');
-Route::put('api/cards/{card_id}/', 'ItemController@create');
-Route::post('api/item/{id}', 'ItemController@update');
-Route::delete('api/item/{id}', 'ItemController@delete');
+Route::controller(CardController::class)->group(function () {
+    Route::put('/api/cards', 'create');
+    Route::delete('/api/cards/{card_id}', 'delete');
+});
+
+Route::controller(ItemController::class)->group(function () {
+    Route::put('/api/cards/{card_id}', 'create');
+    Route::post('/api/item/{id}', 'update');
+    Route::delete('/api/item/{id}', 'delete');
+});
+
 
 // Authentication
-Route::get('login', 'Auth\LoginController@showLoginForm')->name('login');
-Route::post('login', 'Auth\LoginController@login');
-Route::get('logout', 'Auth\LoginController@logout')->name('logout');
-Route::get('register', 'Auth\RegisterController@showRegistrationForm')->name('register');
-Route::post('register', 'Auth\RegisterController@register');
+Route::controller(LoginController::class)->group(function () {
+    Route::get('/login', 'showLoginForm')->name('login');
+    Route::post('/login', 'authenticate');
+    Route::get('/logout', 'logout')->name('logout');
+});
+
+Route::controller(RegisterController::class)->group(function () {
+    Route::get('/register', 'showRegistrationForm')->name('register');
+    Route::post('/register', 'register');
+});
