@@ -2,61 +2,67 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Item;
-use App\Models\Card;
-
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Auth;
+
+use App\Models\Item;
 
 class ItemController extends Controller
 {
-  /**
-   * Creates a new item.
-   *
-   * @param  int  $card_id
-   * @param  Request request containing the description
-   * @return Response
-   */
-  public function create(Request $request, $card_id)
-  {
-    $item = new Item();
-    $item->card_id = $card_id;
-    $this->authorize('create', $item);
-    $item->done = false;
-    $item->description = $request->input('description');
-    $item->save();
-    return $item;
-  }
+    /**
+     * Creates a new item.
+     */
+    public function create(Request $request, $card_id)
+    {
+        // Create a blank new item.
+        $item = new Item();
+
+        // Set item's card.
+        $item->card_id = $card_id;
+
+        // Check if the current user is authorized to create this item.
+        $this->authorize('create', $item);
+
+        // Set item details.
+        $item->done = false;
+        $item->description = $request->input('description');
+
+        // Save the item and return it as JSON.
+        $item->save();
+        return response()->json($item);
+    }
 
     /**
      * Updates the state of an individual item.
-     *
-     * @param  int  $id
-     * @param  Request request containing the new state
-     * @return Response
      */
     public function update(Request $request, $id)
     {
-      $item = Item::find($id);
-      $this->authorize('update', $item);
-      $item->done = $request->input('done');
-      $item->save();
-      return $item;
+        // Find the item.
+        $item = Item::find($id);
+
+        // Check if the current user is authorized to update this item.
+        $this->authorize('update', $item);
+
+        // Update the done property of the item.
+        $item->done = $request->input('done');
+
+        // Save the item and return it as JSON.
+        $item->save();
+        return response()->json($item);
     }
 
     /**
-     * Deletes an individual item.
-     *
-     * @param  int  $id
-     * @return Response
+     * Deletes a specific item.
      */
     public function delete(Request $request, $id)
     {
-      $item = Item::find($id);
-      $this->authorize('delete', $item);
-      $item->delete();
-      return $item;
-    }
+        // Find the item.
+        $item = Item::find($id);
 
+        // Check if the current user is authorized to delete this item.
+        $this->authorize('delete', $item);
+
+        // Delete the item and return it as JSON.
+        $item->delete();
+        return response()->json($item);
+    }
 }
