@@ -2,17 +2,24 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+
+// Added to define Eloquent relationships.
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class Item extends Model
 {
-  // Don't add create and update timestamps in database.
-  public $timestamps  = false;
+    use HasFactory;
 
-  /**
-   * The card this item belongs to.
-   */
-  public function card() {
-    return $this->belongsTo('App\Models\Card');
-  }
+    // Don't add create and update timestamps in database.
+    public $timestamps  = false;
+
+    /**
+     * Get the card where the item is included.
+     */
+    public function card(): BelongsTo
+    {
+        return $this->belongsTo(Card::class);
+    }
 }
