@@ -6,27 +6,40 @@ use App\Models\User;
 use App\Models\Card;
 use App\Models\Item;
 
-use Illuminate\Auth\Access\HandlesAuthorization;
-
 class ItemPolicy
 {
-    use HandlesAuthorization;
-
-    public function create(User $user, Item $item)
+    /**
+     * Create a new policy instance.
+     */
+    public function __construct()
     {
-      // User can only create items in cards they own
-      return $user->id == $item->card->user_id;
+        //
     }
 
-    public function update(User $user, Item $item)
+    /**
+     * Determine if a user can create an item.
+     */
+    public function create(User $user, Item $item): bool
     {
-      // User can only update items in cards they own
-      return $user->id == $item->card->user_id;
+        // User can only create items in cards they own.
+        return $user->id === $item->card->user_id;
     }
 
-    public function delete(User $user, Item $item)
+    /**
+     * Determine if a user can update an item.
+     */
+    public function update(User $user, Item $item): bool
     {
-      // User can only delete items in cards they own
-      return $user->id == $item->card->user_id;
+        // User can only update items in cards they own.
+        return $user->id === $item->card->user_id;
+    }
+
+    /**
+     * Determine if a user can delete an item.
+     */
+    public function delete(User $user, Item $item): bool
+    {
+        // User can only delete items in cards they own.
+        return $user->id === $item->card->user_id;
     }
 }
