@@ -9,27 +9,24 @@ class AppServiceProvider extends ServiceProvider
 {
     /**
      * Register any application services.
-     *
-     * @return void
      */
-    public function register()
+    public function register(): void
     {
         //
     }
 
     /**
      * Bootstrap any application services.
-     *
-     * @return void
      */
-    public function boot()
+    public function boot(): void
     {
-      if(env('FORCE_HTTPS',false)) {
-        error_log('configuring https');
-        $app_url = config("app.url");
-        URL::forceRootUrl($app_url);
-        $schema = explode(':', $app_url)[0];
-        URL::forceScheme($schema);
-      }
+        if(env('FORCE_HTTPS',false)) {
+            error_log('configuring https');
+
+            $app_url = config("app.url");
+            URL::forceRootUrl($app_url);
+            $schema = explode(':', $app_url)[0];
+            URL::forceScheme($schema);
+        }
     }
 }
